requirements.txt
opencv-python>=4.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
pillow-heif>=0.13.0
jsonschema>=4.17.0
pytest>=7.4.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
torch>=2.0.0
torchvision>=0.15.0
segment-anything>=1.0
ultralytics>=8.0.0
rembg>=2.0.0
